Partial commit of z-order parsing
xfl_lexer now parses z-order queries appropriately.

xfl_parser now parses them into an un-runnable ast.

This will cause an error at run time if a z-order queries is entereed in an expression.

The same expression would have failed with an error in the current parser - it is merely that the error is being thrown from a different part of the evaluation tree.