Major commit of z-order and row and column functions
This commit puts in part of the support for:
* z-order queries  (eg /blah/[or(a1, b3)]/bleh/a1)
* row functions    (eg =sum(1:4))
* column functions (eg =sum(a:b))

In this release row and column functions execute and calculate correctly
(but do not recalculate). This calculation is 'pinched out' with an #ERROR!
to mimic the previous behaviour of functions called with this syntax.

z-order functions lex and parse correctly but do not execute. They
then return an #ERROR! exactly as they would do if they failed
to lex or parse