New box and preview behaviour
The behaviour of html.box has now changed. It can take the form
'=html.box.3x4(..)' which is a box three wide by 4 deep.

Height can be anything from 1 to 20 and width from 1 to 12.

There are a couple of other functions that behave similarly:
html.block
html.alert

Also all boxes are now forced to a fixed vertical size