Added proper processing on zseg expressions
The parser post processing of the lexer token stream was a
bit rubbish - it wasn't generating a valid token stream for
itself.

(The lexed and parsed zseg segments should be valid Excel
expressions which when evaluated as {cell, {1, 1}} in a
page context return either true, false or cast to #ERRVAL!)

The expression part of the zseg/seg stream was fine and dandy,
but the path parts were pantsenheimer.

Also cleared a compiled warning.