Robert has a better plan!
Turns out using Excel to generate Erlang terms which can be
converted to Json is plain daft.

Hows about I generate the actual Json and it can be put in the
file system and served as an actual file and even be cached
all over t'internet?

Now would that not have been the way to do it from the start?