Major commit of new functionality.
Previously ranges were expanded during calculation. So a formula like:

=sum(a1:a3) would act as if it was = sum(fetch_range(a1, a2, a3))

If many expressions in different cells  used the same range
then on recalc the full range would be 'refetched' again and again.

This change makes the range an intermediate object which is recalced once
when on of its members changes, and any expressions that use it just use the
intermediate object

However - there is a complication in z-funs (which can also use ranges).

ZExpressions DO NOT cache the range they always recalc from scratch